backend-postgress/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
asyncpg==0.29.0
python-dotenv==1.0.0
python-multipart==0.0.6
bcrypt==4.0.1
PyJWT==2.8.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0
jinja2==3.1.2
aiofiles==23.2.1
alembic==1.12.1
google-generativeai==0.3.1
httpx==0.25.2
